MAINT: Separating MemoryInterface to mixins
